backend/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.9
django-cors-headers==3.10.0
django-extensions==3.1.5
django-seed==0.3.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
Faker==9.8.2
idna==3.3
joblib==1.1.0
krwordrank==1.0.3
numpy==1.21.4
pandas==1.3.4
psycopg2==2.9.2
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
scikit-learn==1.0.1
scipy==1.7.2
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
text-unidecode==1.3
threadpoolctl==3.0.0
toposort==1.7
urllib3==1.26.7
